+++ src/components/layout/pageFooterPMT/PageFooterPMT.tsx
@@ -10,13 +10,17 @@ export default function PageFooterPMT() {
     <Paper sx={{ width: '100vw', position: 'fixed', bottom: 40, left: 0, right: 0 }} elevation={0}>
       <Grid p={4} container direction="row" alignItems="center" justifyContent="space-around">
         <Grid>
-          {(application.content5 as Notification)?.message?.length > 0 && (
-            <TimedAlert
-              color={(application.content5 as Notification)?.level}
-              delay={(application.content5 as Notification)?.delay}
-              testId="timeAlertFooter"
-              text={(application.content5 as Notification)?.message}
-            />
+          {false && ( // Keep until it has been decided if we want to show messages here
+            <Grid>
+              {(application.content5 as Notification)?.message?.length > 0 && (
+                <TimedAlert
+                  color={(application.content5 as Notification)?.level}
+                  delay={(application.content5 as Notification)?.delay}
+                  testId="timeAlertFooter"
+                  text={(application.content5 as Notification)?.message}
+                />
+              )}
+            </Grid>
           )}
         </Grid>
       </Grid>