+++ src/pages/LandingPage/LandingPage.tsx
@@ -26,7 +26,7 @@ import useAxiosAuthClient from '@/services/axios/axiosAuthClient/axiosAuthClient
 import GetProposalAccessForUser from '@/services/axios/get/getProposalAccess/user/getProposalAccessForUser';
 import Proposal from '@/utils/types/proposal';
 import { storeProposalCopy } from '@/utils/storage/proposalData';
-import { validateProposal } from '@/utils/proposalValidation';
+import { validateProposal } from '@/utils/validation/validation';
 import {
   cypressToken,
   DUMMY_PROPOSAL_ID,